Introduce new directory structure
